+++ Source/pushwoosh/example/ios/Flutter/flutter_export_environment.sh
@@ -1,7 +1,7 @@
 #!/bin/sh
 # This is a generated file; do not edit or check into version control.
-export "FLUTTER_ROOT=/Users/vromanychev/flutter"
-export "FLUTTER_APPLICATION_PATH=/Users/vromanychev/Desktop/projects/Flutter/flutter-public/Source/pushwoosh/example"
+export "FLUTTER_ROOT=/Users/vromanychev/development/flutter"
+export "FLUTTER_APPLICATION_PATH=/Users/vromanychev/Desktop/projects/flutter/Sources/pushwoosh/example"
 export "COCOAPODS_PARALLEL_CODE_SIGN=true"
 export "FLUTTER_TARGET=lib/main.dart"
 export "FLUTTER_BUILD_DIR=build"